controllers and services

+++ ClimateChecker/src/main/java/Ontdekstation013/ClimateChecker/models/dto/sensorDto.java
@@ -7,9 +7,10 @@ import lombok.Setter;
 @Setter
 public class sensorDto extends Dto{
 
-    int id;
-    int stationId;
-    int typeId;
+
+    long id;
+    long stationId;
+    long typeId;
     float data;
 
 }
